default.nix: packageLockFile: Initial revision
Given the package info JSON generated by buildroot, generate a
JSON-formatted lock file that specifies which files need to be
downloaded, which URIs to use, and what the resulting hashes should
be.

I should be able to consume this lock file in nix to create
fixed-output derivations for each input file, as well as a "downloads"
directory to use as an input to the buildroot build.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>