default.nix: packageInputs: Initial revision
Buildroot requires a bunch of source tarballs that it will download
from the internet if they don't alreay exist. Have nix pre-download
these since the builder cannot access the internet during the actual
build.

This definition is generated from a lockfile, which can be built for
the given defconfig/source tree using the existing packageLockFile
derivation. This can't be generated dynamically since nix is a
non-monadic build system. Instead, the lock file should be built and
checked into VCS, then passed as an input to the build. A future check
will ensure that the lockfile is up-to-date.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>