default.nix: buildrootBase: Patch shebangs in all support scripts
There are a bunch of scripts in support/{scripts,download} that are
used in the build but hardcode /usr/bin/env as the shebang. Patch
these to point at env in the nix store.

Note that we don't want to patch shebangs globally, since many scripts
may end up on-device and we don't want store paths to propagate into
the rootfs.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>