default.nix: Use a FHS environment instead of patching scripts
There are way too many places in buildroot that hardcode /bin or
/usr/bin, some of which are in unpacked tarballs (we can't naively
patch those...). Instead of playing whack-a-mole with everything that
comes up (including testing with all packages enabled), just build
everything in a Filesystem Hierarchy Standard ("FHS") environment.

An added bonus of this approach is that it is much less likely that
stray references to the store leak into the resulting buildroot image.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>