default.nix: buildrootBase: Disable setuid in util-linux
During the host util-linux build, buildroot attempts to set the setuid
bit on mount (which allows mounting as non-root users). Since we are
in a user namespace running as UID 0, this fails in the nix sandbox:

chmod 4755 /build/source/output/host/bin/mount
chmod: changing permissions of '/build/source/output/host/bin/mount': Operation not permitted

Setting the setuid bit is not necessary since buildroot knows it is
not running as root and will not try to mount things. Work around this
by patching out the configure flag.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>