default.nix: buildroot: Initial revision
Add the initial implementation of the buildroot builder as a
'buildroot' output for the flake. This builder symlinks the
pre-downloaded source tarballs into the download directory then calls
'make' inside the FHS environment.

All of the files in the resulting 'images' directory are copied to the
output; the contents of this directory may change based on the
defconfig, so I don't want to hardcode any specific file outputs.

I may make this customizable or copy some other artifacts (debug
symbols?) in the future.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>