default.nix: mkBuildroot: Add nativeBuildInputs argument
Similar to the nativeBuildInputs argument to mkDerivation, this
argument adds additional native (i.e. host) packages to the
environment used during the build.

This is useful for adding additional tools and libraries that are only
used in a subset of buildroot packages, such as additional downloaders
listed in DL_TOOLS_DEPENDENCIES like git and svn.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>