default.nix: lockedInputs: Fix files with query parameters
When installing symlinks for the lockedInputs build, files may contain
query parameters including '=', which breaks the symlink command for
that file. This would result in some downloaded outputs (such as
spidev_test.c) not being present in $BR2_DL_DIR, so the build would
fail trying to download it.

Fix this by adding additional quoting. Note that nix fixes this for
the input store path automatically, but not for the file reference.

Fixes: 0e58b98 ("default.nix: packageInputs: Initial revision")
Signed-off-by: Brian Kubisiak <brian@kubisiak.com>