default.nix: buildroot: Add sdk output
The entire compiler toolchain must already get built during the
buildroot build, so building and installing the sdk adds minimal time
when building the derivation.

Note that it isn't entirely usable due to bad shebangs on nix or guix
systems; I'll need to go through the fixup phase and re-enable steps
that are required here. This should be usable on most linux distros
as-is.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>