default.nix: Allow passing a path as a defconfig
When the defconfig is not located at ${src}/configs/*_defconfig
(i.e. it is an absolute path), the configure phase must invoke make
using the BR2_DEFCONFIG environment variable instead.

When the defconfig is passed as a path, use this alternate form and
pass in the path to the config in the store.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>